Refactor app structure: move App component and routes to separate files, convert default exports to named exports for better modularity
