+++ src/app/index.ts
@@ -1,0 +1,2 @@
+export { App } from './app';
+export { Routes } from './routes';
