+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './app';
+import { App } from '@/app';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
