+++ src/pages/favorites-empty/favorites-empty.tsx
@@ -20,4 +20,4 @@ function FavoritesEmpty() {
   );
 }
 
-export default FavoritesEmpty;
+export { FavoritesEmpty };
