+++ src/pages/index.ts
@@ -1,0 +1,9 @@
+export { Favorites } from './favorites/favorites';
+export { FavoritesEmpty } from './favorites-empty/favorites-empty';
+export { Login } from './login/login';
+export { Main } from './main/main';
+export { MainEmpty } from './main-empty/main-empty';
+export { NotFound } from './not-found/not-found';
+export { Offer } from './offer/offer';
+export { OfferNotLogged } from './offer-not-logged/offer-not-logged';
+export { PrivateRoute } from './private-route/private-route';
