+++ src/pages/login/login.tsx
@@ -65,4 +65,4 @@ function Login() {
   );
 }
 
-export default Login;
+export { Login };
