+++ src/pages/not-found/not-found.tsx
@@ -22,4 +22,4 @@ function NotFound() {
   );
 }
 
-export default NotFound;
+export { NotFound };
